add ProfilesLogic and migrated usages of ProfilesDb from AccountsLogic to it
